Add VBPI-vs-MCMC comparison notebook
examples/vbpi_vs_mcmc_comparison.ipynb simulates a small rooted 5-taxon tree
under Jukes-Cantor, then infers the topology posterior two ways over the same
joint p(T, b | D):

- VBPI with the enumerated subsplit Bayesian network (complete support over all
  105 rooted topologies) + per-split log-normal branches, trained with VIMCO;
- the rooted-NNI Metropolis-Hastings sampler.

It reads off q(T) for every topology, compares it to the MCMC visiting
frequencies (per-rooted scatter, top-topology bars, and an aggregation to
unrooted topologies that removes the JC rooting non-identifiability), and marks
the true tree. In the executed run both methods place ~0.13 on the true rooted
topology and >0.86 on the true unrooted tree, spreading mass similarly across the
exchangeable rootings. Everything runs in rooted-tree space on treeflow's JC
model, native likelihood op, and TFP TransitionKernel MCMC.

Also move the shared VBPI test helpers into test/vbpi/vbpi_test_helpers.py so
the test files import them under a unique name rather than via `conftest`, which
collided with the native suite's conftest when both were collected together.
